refactored queuereq method
