+++ routes/upload.js
@@ -1,6 +1,7 @@
 var utility = require('../utility');
 var domain = require('domain');
-var queueReq = utility.queueReq;
+var operations = require('../operations');
+var queueReq = operations.queueReq;
 var multiparty = require('multiparty');
 var express = require('express');
 var upload = express.Router();
